// src/types/Requisition.ts
export type RequisitionInfo = {
  requisitionId: string;
  clientId: number;
  requesterId: string;
  department: string;
  testType: string;
  testCode: string;
  clinicalInfo: string;
  dateRequested: string;
  priority: string;
  status: string;
  labName: string;
  labAddress: string;
  resultDate: string;
  notes: string;
}